requirements.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.9.0
click==8.1.7
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
packaging==24.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
redis==5.0.3
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.1
